Put bus interaction inside constraint_system
