+++ constraint-solver/src/lib.rs
@@ -1,6 +1,5 @@
 //! Tooling used for analysis and solving of constraints.
 
-pub mod bus_interaction;
 pub mod constraint_system;
 pub mod effect;
 pub mod quadratic_symbolic_expression;
